fix(clickhouse): the duration column holds a float
SHAPE-1 made `duration_ms` a float with microsecond precision. The recommended
DDL still declared `UInt64`, so ClickHouse refused every row with a fractional
duration. The drain could not insert its own event, and `make integration`
failed.

The column is now `Float64`, which holds the value the event carries.

+++ drain/clickhouse/ddl.go
@@ -11,7 +11,7 @@ func DDL(database, table string) string {
     timestamp DateTime64(9),
     level LowCardinality(String),
     operation String,
-    duration_ms UInt64,
+    duration_ms Float64,
     outcome LowCardinality(String),
     service_name LowCardinality(String),
     service_version String,
